dyn: use unsigned instead of size_t in algorithm signatures
size_t is really for size of objects in bytes. The main reason,
however, is to avoid having too many different types to deal with in
dyn.

* vcsn/dyn/algos.hh (enumerate, chain): Use unsigned instead of size_t.
* lib/vcsn/algos/concatenate.cc, lib/vcsn/algos/enumerate.cc,
* vcsn/algos/concatenate.hh, vcsn/algos/enumerate.hh: Adjust.

+++ lib/vcsn/algos/concatenate.cc
@@ -17,7 +17,7 @@ namespace vcsn
     REGISTER_DEFINE(chain);
 
     automaton
-    chain(const automaton& aut, size_t n)
+    chain(const automaton& aut, unsigned n)
     {
       return detail::chain_registry().call(aut->vname(false), aut, n);
     }
